updating changes related to project
