+++ .env.example
@@ -39,3 +39,7 @@ SECRET_IV=
 NEXT_PUBLIC_SECRET_KEY=
 NEXT_PUBLIC_TOKEN_MINT_ADDRESS=
 NEXT_PUBLIC_SHDW_DRIVE_PUB_KEY=
+
+
+NEXT_PUBLIC_PTVB_TOKEN=
+NEXT_PUBLIC_PTVR_TOKEN=
